Remove optimizer f_ methods
Arbitrary torch exceptions in the middle of performing an optimizer step
might leave the optimizer or parameters in a bad state.

Instead have the main methods return Result<..., OptimizerStepError>
so that implementations can be more careful about cleaning up exceptions
(instead of f_foo()? everywhere) and the caller can choose what to do
with the error.